Combined all existing c2cpg data-flow tests
Added test for https://github.com/joernio/joern/issues/892
Added test for https://github.com/joernio/joern/issues/913
