+++ joern-cli/frontends/c2cpg/src/test/scala/io/joern/c2cpg/testfixtures/DataFlowCodeToCpgSuite.scala
@@ -7,18 +7,16 @@ import io.joern.dataflowengineoss.layers.dataflows.{OssDataFlow, OssDataFlowOpti
 import io.joern.dataflowengineoss.queryengine.EngineContext
 import io.joern.dataflowengineoss.semanticsloader.{Parser, Semantics}
 import io.shiftleft.semanticcpg.language._
-import io.shiftleft.semanticcpg.language.dotextension.ImageViewer
 import io.shiftleft.semanticcpg.layers.{Base, CallGraph, ControlFlow, LayerCreatorContext, TypeRelations}
 import io.shiftleft.utils.ProjectRoot
-import overflowdb.traversal.Traversal
-
-import scala.sys.process.Process
-import scala.util.Try
 
 class DataFlowCodeToCpgSuite extends CCodeToCpgSuite {
 
-  var semanticsFilename = ProjectRoot.relativise("dataflowengineoss/src/test/resources/default.semantics")
+  var semanticsFilename: String =
+    ProjectRoot.relativise("dataflowengineoss/src/test/resources/default.semantics")
+
   var semantics: Semantics = _
+
   implicit var context: EngineContext = _
 
   override def beforeAll(): Unit = {
@@ -26,11 +24,6 @@ class DataFlowCodeToCpgSuite extends CCodeToCpgSuite {
     semantics = Semantics.fromList(new Parser().parseFile(semanticsFilename))
     context = EngineContext(semantics)
   }
-
-  implicit val viewer: ImageViewer = (pathStr: String) =>
-    Try {
-      Process(Seq("xdg-open", pathStr)).!!
-    }
 
   override def passes(cpg: Cpg): Unit = {
     val context = new LayerCreatorContext(cpg)
@@ -43,17 +36,7 @@ class DataFlowCodeToCpgSuite extends CCodeToCpgSuite {
     new OssDataFlow(options).run(context)
   }
 
-  protected implicit def int2IntegerOption(x: Int): Option[Integer] =
-    Some(x)
-
-  protected def getMemberOfType(cpg: Cpg, typeName: String, memberName: String): Traversal[Member] =
-    cpg.typeDecl.nameExact(typeName).member.nameExact(memberName)
-
-  protected def getMethodOfType(cpg: Cpg, typeName: String, methodName: String): Traversal[Method] =
-    cpg.typeDecl.nameExact(typeName).method.nameExact(methodName)
-
-  protected def getLiteralOfType(cpg: Cpg, typeName: String, literalName: String): Traversal[Literal] =
-    cpg.typeDecl.nameExact(typeName).method.isLiteral.codeExact(literalName)
+  protected implicit def int2IntegerOption(x: Int): Option[Integer] = Some(x)
 
   protected def flowToResultPairs(path: Path): List[(String, Option[Integer])] = {
     val pairs = path.elements.map {
