+++ querydb/src/test/scala/io/joern/suites/CQueryTestSuite.scala
@@ -1,10 +1,9 @@
 package io.joern.suites
 
 import io.joern.util.QueryUtil
-import io.joern.util.QueryUtil.allQueries
 import io.shiftleft.codepropertygraph.generated.nodes
 import io.joern.console.scan._
-import io.joern.console.{QueryBundle, QueryDatabase}
+import io.joern.console.QueryBundle
 import io.joern.console.Query
 import io.joern.c2cpg.testfixtures.DataFlowCodeToCpgSuite
 import io.shiftleft.semanticcpg.language._
